added cstint types + more

+++ Chess-Engine/c++/src/precompute.cpp
@@ -14,7 +14,7 @@ uint64_t king_moveboards[64];
 uint64_t rook_relevant_occupancy[64];
 uint64_t bishop_relevant_occupancy[64];
 
-uint_fast16_t rook_offsets[64];
+uint_fast32_t rook_offsets[64];
 uint_fast16_t bishop_offsets[64];
 
 uint64_t not_A_file = 0x7F7F7F7F7F7F7F7F;
@@ -29,10 +29,10 @@ void initLookupTables() {
 }
 
 void initRookLookups() {
-    uint_fast8_t offset = 0;
+    int_fast32_t offset = 0;
     uint64_t subset = 0;
 
-    for (uint_fast8_t sq = 0; sq <= 63; sq++) {
+    for (int_fast8_t sq = 0; sq <= 63; sq++) {
         rook_relevant_occupancy[sq] = relevantOccupancyRook(sq);
         rook_offsets[sq] = offset;
 
@@ -48,10 +48,10 @@ void initRookLookups() {
 }
 
 void initBishopLookups() {
-    uint_fast8_t offset = 0;
+    int_fast16_t offset = 0;
     uint64_t subset = 0;
 
-    for (uint_fast8_t sq = 0; sq <= 63; sq++) {
+    for (int_fast8_t sq = 0; sq <= 63; sq++) {
         bishop_relevant_occupancy[sq] = relevantOccupancyBishop(sq);
         bishop_offsets[sq] = offset;
 
@@ -70,7 +70,7 @@ void initKnightLookups() {
     uint64_t col;
     uint64_t moves; 
 
-    for (uint_fast8_t sq = 0; sq <= 63; sq++) {
+    for (int_fast8_t sq = 0; sq <= 63; sq++) {
         moves = 0;
         zeroAndSetBit(pos, sq);
         col = (sq % 8);
@@ -103,7 +103,7 @@ void initPawnLookups() {
     uint64_t moves;
     uint64_t takes;
 
-    for (uint_fast8_t sq = 8; sq <= 55; sq++) {
+    for (int_fast8_t sq = 8; sq <= 55; sq++) {
         moves = 0;
         takes = 0;
         zeroAndSetBit(pos, sq);
@@ -118,7 +118,7 @@ void initPawnLookups() {
         pawn_moveboards[white][pushes][sq] = moves;
     }
 
-    for (uint_fast8_t sq = 55; sq >= 8; sq--) {
+    for (int_fast8_t sq = 55; sq >= 8; sq--) {
         moves = 0;
         takes = 0;
         zeroAndSetBit(pos, sq);
@@ -138,7 +138,7 @@ void initKingLookups() {
     uint64_t moves;
     uint64_t pos;
 
-    for (uint_fast8_t sq = 0; sq <= 63; sq++) {
+    for (int_fast8_t sq = 0; sq <= 63; sq++) {
         moves = 0;
         zeroAndSetBit(pos, sq);
